+++ src/ray_casting/ray_length.c
@@ -6,13 +6,13 @@
 /*   By: alida-si <alida-si@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 11:35:58 by alida-si          #+#    #+#             */
-/*   Updated: 2023/03/30 09:42:13 by alida-si         ###   ########.fr       */
+/*   Updated: 2023/03/30 13:32:44 by alida-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-int	distance_to_horizontal_wall(t_game *game, float angle)
+float	distance_to_horizontal_wall(t_game *game, float angle)
 {
 	if (looking_up(game->player->angle))
 		return (distance_to_horizontal_wall_up(game, angle));
@@ -20,7 +20,7 @@ int	distance_to_horizontal_wall(t_game *game, float angle)
 		return (distance_to_horizontal_wall_down(game, angle));
 }
 
-int	distance_to_vertical_wall(t_game *game, float angle)
+float	distance_to_vertical_wall(t_game *game, float angle)
 {
 	if (looking_right(game->player->angle))
 		return (distance_to_vertical_wall_right(game, angle));
@@ -28,7 +28,7 @@ int	distance_to_vertical_wall(t_game *game, float angle)
 		return (distance_to_vertical_wall_left(game, angle));
 }
 
-int	find_wall_distance(t_game *game, float angle)
+float	find_wall_distance(t_game *game, float angle)
 {
 	int	vertical_wall_distance;
 	int	horizontal_wall_distance;
